Trying push notification for match

+++ api/deals/api.py
@@ -1,8 +1,13 @@
 import json
+import requests
 from django.http import HttpResponse
 from .models import Deals, ClientUsers, BlackLists, Choices, Requests
 from .serializers import DealsSerializer, BlackListsSerializer, ClientUsersSerializer, ChoicesSerializer, RequestsSerializer
 
+
+FCM_ENDPOINT = 'https://fcm.googleapis.com/fcm/send'
+FCM_KEY = 'AAAAzfk1l5E:APA91bFnzxOs2ne4MKnQM4qAKqBNE6UjiQqOA3og3Ud4xBlZU0KL9OJUDuzFdf6Suvy7lqd-hlxsYGPXUJjCS7qSHstixiCes1mDhpLY8-ZsykRkHQwVYtUaktGagNrd1AUvRFIq9usF'
+FCM_USER = '884649334673'
 def getAllDeals(request):
     queryset = Deals.objects.all()
     serializer = DealsSerializer(queryset, many=True)
@@ -76,6 +81,7 @@ def addRequest(request, uid, dealid, c):
         for i in l:
             choice = Choices.objects.get(pk=int(i))
             request.choices.add(choice)
+        matchTrigger()
         return HttpResponse('<H1>SUCCESS</H1>')
     except Exception as e:
         return HttpResponse('<H1>%s</H1>' %str(e))
@@ -95,9 +101,10 @@ def getRequestById(request, uid):
     queryset = Requests.objects.filter(clientuser=clientuser)
     serializer = RequestsSerializer(queryset, many=True)
     return HttpResponse(json.dumps(serializer.data))
-'''
-def matchTrigger(request):
-    queryset = Request.objects.all()
+
+
+def matchTrigger():
+    queryset = Requests.objects.all()
     queueidRecent = queryset[len(queryset)-1]['id']
     uidRecent = queryset[len(queryset)-1]['clientuser']
     dealRecent = queryset[len(queryset)-1]['deal']
@@ -106,14 +113,47 @@ def matchTrigger(request):
     for i in range(0,len(queryset)-1):
         if(queryset[i]['deal'] == dealRecent):
             if(set(choicesRecent) & set(queryset[i]['choices'])):
-                matchDict = {}
-                matchDict['uid1'] = uidRecent
-                matchDict['uid2'] = queryset[i]['clientuser']
-                matchDict['deal'] = dealRecent
-                matchDict['locations'] = list(set(choicesRecent) & set(queryset[i]['choices']))
-                #remove both users from database
-                SomeModel.objects.filter(id=queueidRecent).delete()
-                SomeModel.objects.filter(id=queryset[i]['id']).delete()
-                #send push-notification to both users
-                return matchDict
-'''
+                clientuser1 = ClientUsers.objects.get(uid=uidRecent)
+                clientuser2 = ClientUsers.objects.get(uid=queryset[i]['clientuser'])
+                deal = Deals.objects.get(dealid=dealRecent)
+                if(not(BlackLists.objects.get(clientuser1=clientuser1, clientuser2=clientuser2, deal=deal) 
+                or BlackLists.objects.get(clientuser1=clientuser2, clientuser2=clientuser1, deal=deal))):
+                    data1 = {}
+                    body1 = []
+                    body1.append(json.dumps(ClientUsersSerializer(clientuser1, many=False)))
+                    body1.append(json.dumps(ClientUsersSerializer(clientuser2, many=False)))
+                    body1.append(json.dumps(DealsSerializer(deal, many=False)))
+                    data1['notification'] = {}
+                    data1['notification']['title'] = "Found a match!"
+                    data1['notification']['body'] = body1
+                    data1['notification']['icon'] = "firebase-logo.png"
+                    data1['notification']['click_action'] = "http://localhost:8081"
+                    data1['to'] = clientuser1.token
+
+                    data2 = {}
+                    body2 = []
+                    body2.append(json.dumps(ClientUsersSerializer(clientuser1, many=False)))
+                    body2.append(json.dumps(ClientUsersSerializer(clientuser2, many=False)))
+                    body2.append(json.dumps(DealsSerializer(deal, many=False)))
+                    data2['notification'] = {}
+                    data2['notification']['title'] = "Found a match!"
+                    data2['notification']['body'] = body2
+                    data2['notification']['icon'] = "firebase-logo.png"
+                    data2['notification']['click_action'] = "http://localhost:8081"
+                    data2['to'] = clientuser2.token
+
+                    requests.post(FCM_ENDPOINT, data = data1, auth=(FCM_USER, FCM_KEY))
+                    requests.post(FCM_ENDPOINT, data = data2, auth=(FCM_USER, FCM_KEY))
+
+                    '''
+                    matchDict = {}
+                    matchDict['uid1'] = uidRecent
+                    matchDict['uid2'] = queryset[i]['clientuser']
+                    matchDict['deal'] = dealRecent
+                    matchDict['locations'] = list(set(choicesRecent) & set(queryset[i]['choices']))
+                    #remove both users from database
+                    SomeModel.objects.filter(id=queueidRecent).delete()
+                    SomeModel.objects.filter(id=queryset[i]['id']).delete()
+                    #send push-notification to both users
+                    return matchDict
+                    '''
